address code review issues: LRU eviction, extract chat-cache utility, blob URL leak, deduplicate CSS
- Fix cache eviction in video-cache.js and pdf-cache.js to use getAll()
  + reduce by timestamp instead of lexicographic cursor order (real LRU)
- Move IndexedDB helpers from MessageBubble.vue to src/utils/chat-cache.js
  and update imports in actions.js and profile.vue (architectural fix)
- Add max-entries eviction with URL.revokeObjectURL to video-blob-cache.js
- Move duplicate .q-dialog__inner--maximized > div override to app.sass
