+++ src/components/chat/MessageBubble.vue
@@ -235,160 +235,13 @@ import { getCachedVideoBlob } from 'src/utils/video-blob-cache'
 import { getCachedVideo, setCachedVideo, getCachedVideoThumb, setCachedVideoThumb } from 'src/utils/video-cache'
 import PdfViewerDialog from 'src/components/chat/PdfViewerDialog.vue'
 import { getCachedPdf, setCachedPdf } from 'src/utils/pdf-cache'
-
-const _replyThumbnailCache = new Map()
-
-// Module-level cache for image thumbnails (LRU with max 100 entries)
-const _imageThumbnailCache = new Map()
-const MAX_THUMBNAIL_CACHE_SIZE = 200
-
-// IndexedDB for persistent thumbnail cache
-const DB_NAME = 'paytaca-chat-cache'
-const DB_VERSION = 3
-const STORE_NAME = 'thumbnails'
-
-let _dbPromise = null
-
-function openDatabase() {
-  if (!_dbPromise) {
-    _dbPromise = new Promise((resolve, reject) => {
-      const request = indexedDB.open(DB_NAME, DB_VERSION)
-      
-      request.onerror = () => reject(request.error)
-      request.onsuccess = () => resolve(request.result)
-      
-      request.onupgradeneeded = (event) => {
-        const db = event.target.result
-        if (!db.objectStoreNames.contains(STORE_NAME)) {
-          db.createObjectStore(STORE_NAME, { keyPath: 'id' })
-        }
-        if (!db.objectStoreNames.contains('videos')) {
-          db.createObjectStore('videos', { keyPath: 'id' })
-        }
-        if (!db.objectStoreNames.contains('videoThumbs')) {
-          db.createObjectStore('videoThumbs', { keyPath: 'id' })
-        }
-        if (!db.objectStoreNames.contains('pdfs')) {
-          db.createObjectStore('pdfs', { keyPath: 'id' })
-        }
-      }
-    })
-  }
-  return _dbPromise
-}
-
-async function getThumbnailFromDB(cacheKey) {
-  try {
-    const db = await openDatabase()
-    return new Promise((resolve) => {
-      const tx = db.transaction(STORE_NAME, 'readonly')
-      const store = tx.objectStore(STORE_NAME)
-      const request = store.get(cacheKey)
-      request.onsuccess = () => resolve(request.result?.thumbnailUrl || null)
-      request.onerror = () => resolve(null)
-    })
-  } catch {
-    return null
-  }
-}
-
-async function saveThumbnailToDB(cacheKey, thumbnailUrl) {
-  try {
-    const db = await openDatabase()
-    const tx = db.transaction(STORE_NAME, 'readwrite')
-    const store = tx.objectStore(STORE_NAME)
-    store.put({ id: cacheKey, thumbnailUrl, timestamp: Date.now() })
-  } catch (err) {
-    console.warn('Failed to save thumbnail to IndexedDB:', err)
-  }
-}
-
-export async function clearChatCache() {
-  try {
-    const db = await openDatabase()
-    const tx = db.transaction([STORE_NAME, 'videos', 'videoThumbs', 'pdfs'], 'readwrite')
-    tx.objectStore(STORE_NAME).clear()
-    tx.objectStore('videos').clear()
-    tx.objectStore('videoThumbs').clear()
-    tx.objectStore('pdfs').clear()
-    _imageThumbnailCache.clear()
-    _replyThumbnailCache.clear()
-    return true
-  } catch (err) {
-    console.error('Failed to clear chat cache:', err)
-    return false
-  }
-}
-
-export async function hasChatCache() {
-  try {
-    if (_imageThumbnailCache.size > 0 || _replyThumbnailCache.size > 0) return true
-    const db = await openDatabase()
-    const stores = [STORE_NAME, 'videos', 'videoThumbs', 'pdfs']
-    for (const name of stores) {
-      if (!db.objectStoreNames.contains(name)) continue
-      const has = await new Promise((resolve) => {
-        const tx = db.transaction(name, 'readonly')
-        const store = tx.objectStore(name)
-        const countReq = store.count()
-        countReq.onsuccess = () => resolve(countReq.result > 0)
-        countReq.onerror = () => resolve(false)
-      })
-      if (has) return true
-    }
-    return false
-  } catch {
-    return false
-  }
-}
-
-export async function getChatCacheSize () {
-  try {
-    let totalBytes = 0
-    for (const url of _imageThumbnailCache.values()) {
-      totalBytes += typeof url === 'string' ? url.length * 0.75 : 0
-    }
-    for (const url of _replyThumbnailCache.values()) {
-      totalBytes += typeof url === 'string' ? url.length * 0.75 : 0
-    }
-    const db = await openDatabase()
-    const stores = [STORE_NAME, 'videos', 'videoThumbs', 'pdfs']
-    for (const name of stores) {
-      if (!db.objectStoreNames.contains(name)) continue
-      await new Promise((resolve) => {
-        const tx = db.transaction(name, 'readonly')
-        const store = tx.objectStore(name)
-        const cursorReq = store.openCursor()
-        cursorReq.onsuccess = () => {
-          const cursor = cursorReq.result
-          if (cursor) {
-            const val = cursor.value
-            if (val?.thumbnailUrl) totalBytes += val.thumbnailUrl.length * 0.75
-            if (val?.blob?.size) totalBytes += val.blob.size
-            cursor.continue()
-          } else {
-            resolve()
-          }
-        }
-        cursorReq.onerror = () => resolve()
-      })
-    }
-    return totalBytes
-  } catch {
-    return 0
-  }
-}
-
-function evictOldestThumbnail() {
-  if (_imageThumbnailCache.size >= MAX_THUMBNAIL_CACHE_SIZE) {
-    const firstKey = _imageThumbnailCache.keys().next().value
-    const url = _imageThumbnailCache.get(firstKey)
-    if (url && url.startsWith('blob:')) {
-      URL.revokeObjectURL(url)
-    }
-    _imageThumbnailCache.delete(firstKey)
-  }
-}
+import {
+  _imageThumbnailCache,
+  _replyThumbnailCache,
+  evictOldestThumbnail,
+  getThumbnailFromDB,
+  saveThumbnailToDB,
+} from 'src/utils/chat-cache'
 
 function bytesToHex(bytes) {
   return Array.from(bytes).map(b => b.toString(16).padStart(2, '0')).join('')
@@ -1957,11 +1810,6 @@ export default {
 </style>
 
 <style>
-/* Fullscreen dialogs must cover safe areas on notched phones */
-.q-dialog__inner--maximized > div {
-  max-height: 100dvh !important;
-}
-
 /* Global styles for seen-by q-menu portal content (unscoped — q-menu renders outside component) */
 .seen-by-menu-inner {
   padding: 10px 14px;
