+++ src/pages/apps/chat/profile.vue
@@ -342,7 +342,7 @@ import { validateAddress } from 'src/utils/send-page-utils'
 import { getWalletByNetwork } from 'src/wallet/chipnet'
 import { cachedLoadWallet } from 'src/wallet'
 import { npubEncode } from 'nostr-tools/nip19'
-import { clearChatCache, hasChatCache, getChatCacheSize } from 'src/components/chat/MessageBubble.vue'
+import { clearChatCache, hasChatCache, getChatCacheSize } from 'src/utils/chat-cache'
 import { copyToClipboard } from 'quasar'
 import { uploadPublicToBlossom } from 'src/wallet/nostr-media'
 import ResetChatDialog from './ResetChatDialog.vue'
