+++ src/utils/pdf-cache.js
@@ -65,13 +65,12 @@ export async function setCachedPdf (id, blob, mimeType) {
     const countReq = store.count()
     countReq.onsuccess = () => {
       if (countReq.result >= MAX_PDF_ENTRIES) {
-        const cursorReq = store.openCursor()
-        cursorReq.onsuccess = () => {
-          const cursor = cursorReq.result
-          if (cursor) {
-            cursor.delete()
-            store.put({ id, blob, mimeType, timestamp: Date.now() })
-          }
+        const allReq = store.getAll()
+        allReq.onsuccess = () => {
+          const entries = allReq.result
+          const oldest = entries.reduce((a, b) => a.timestamp < b.timestamp ? a : b)
+          store.delete(oldest.id)
+          store.put({ id, blob, mimeType, timestamp: Date.now() })
         }
       } else {
         store.put({ id, blob, mimeType, timestamp: Date.now() })
